+++ javaslang/src/main/java/javaslang/collection/Traversable.java
@@ -5,12 +5,19 @@
  */
 package javaslang.collection;
 
-import javaslang.*;
+import javaslang.Tuple2;
+import javaslang.Tuple3;
+import javaslang.Value;
 import javaslang.control.Option;
 
 import java.math.BigInteger;
-import java.util.*;
-import java.util.function.*;
+import java.util.Comparator;
+import java.util.NoSuchElementException;
+import java.util.Objects;
+import java.util.function.BiFunction;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.Predicate;
 
 /**
  * An interface for inherently recursive, multi-valued data structures. The order of elements is determined by
@@ -38,10 +45,10 @@ import java.util.function.*;
  * Iteration:
  *
  * <ul>
- * <li>{@link #grouped(long)}</li>
+ * <li>{@link #grouped(int)}</li>
  * <li>{@link #iterator()}</li>
- * <li>{@link #sliding(long)}</li>
- * <li>{@link #sliding(long, long)}</li>
+ * <li>{@link #sliding(int)}</li>
+ * <li>{@link #sliding(int, int)}</li>
  * </ul>
  *
  * Numeric operations:
@@ -79,8 +86,8 @@ import java.util.function.*;
  * Selection:
  *
  * <ul>
- * <li>{@link #drop(long)}</li>
- * <li>{@link #dropRight(long)}</li>
+ * <li>{@link #drop(int)}</li>
+ * <li>{@link #dropRight(int)}</li>
  * <li>{@link #dropUntil(Predicate)}</li>
  * <li>{@link #dropWhile(Predicate)}</li>
  * <li>{@link #filter(Predicate)}</li>
@@ -89,8 +96,8 @@ import java.util.function.*;
  * <li>{@link #groupBy(Function)}</li>
  * <li>{@link #partition(Predicate)}</li>
  * <li>{@link #retainAll(Iterable)}</li>
- * <li>{@link #take(long)}</li>
- * <li>{@link #takeRight(long)}</li>
+ * <li>{@link #take(int)}</li>
+ * <li>{@link #takeRight(int)}</li>
  * <li>{@link #takeUntil(Predicate)}</li>
  * <li>{@link #takeWhile(Predicate)}</li>
  * </ul>
@@ -255,7 +262,7 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
      * @return a new instance consisting of all elements of this except the first n ones, or else the empty instance,
      * if this has less than n elements.
      */
-    Traversable<T> drop(long n);
+    Traversable<T> drop(int n);
 
     /**
      * Drops the last n elements of this or all elements, if this length &lt; n.
@@ -264,7 +271,7 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
      * @return a new instance consisting of all elements of this except the last n ones, or else the empty instance,
      * if this has less than n elements.
      */
-    Traversable<T> dropRight(long n);
+    Traversable<T> dropRight(int n);
 
     /**
      * Drops elements until the predicate holds for the current element.
@@ -467,14 +474,14 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
      * </code>
      * </pre>
      *
-     * Please note that {@code grouped(int)} is a special case of {@linkplain #sliding(long, long)}, i.e.
+     * Please note that {@code grouped(int)} is a special case of {@linkplain #sliding(int, int)}, i.e.
      * {@code grouped(size)} is the same as {@code sliding(size, size)}.
      *
      * @param size a positive block size
      * @return A new Iterator of grouped blocks of the given size
      * @throws IllegalArgumentException if {@code size} is negative or zero
      */
-    Iterator<? extends Traversable<T>> grouped(long size);
+    Iterator<? extends Traversable<T>> grouped(int size);
 
     /**
      * Checks if this Traversable is known to have a finite size.
@@ -1029,13 +1036,13 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
 
     /**
      * Slides a window of a specific {@code size} and step size 1 over this {@code Traversable} by calling
-     * {@link #sliding(long, long)}.
+     * {@link #sliding(int, int)}.
      *
      * @param size a positive window size
      * @return a new Iterator of windows of a specific size using step size 1
      * @throws IllegalArgumentException if {@code size} is negative or zero
      */
-    Iterator<? extends Traversable<T>> sliding(long size);
+    Iterator<? extends Traversable<T>> sliding(int size);
 
     /**
      * Slides a window of a specific {@code size} and {@code step} size over this {@code Traversable}.
@@ -1056,7 +1063,7 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
      * @return a new Iterator of windows of a specific size using a specific step size
      * @throws IllegalArgumentException if {@code size} or {@code step} are negative or zero
      */
-    Iterator<? extends Traversable<T>> sliding(long size, long step);
+    Iterator<? extends Traversable<T>> sliding(int size, int step);
 
     /**
      * Returns a tuple where the first element is the longest prefix of elements that satisfy the given
@@ -1132,7 +1139,7 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
      * @param n The number of elements to take.
      * @return A new instance consisting the first n elements of this or all elements, if this has less than n elements.
      */
-    Traversable<T> take(long n);
+    Traversable<T> take(int n);
 
     /**
      * Takes the last n elements of this or all elements, if this length &lt; n.
@@ -1145,7 +1152,7 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
      * @param n The number of elements to take.
      * @return A new instance consisting the first n elements of this or all elements, if this has less than n elements.
      */
-    Traversable<T> takeRight(long n);
+    Traversable<T> takeRight(int n);
 
     /**
      * Takes elements until the predicate holds for the current element.
@@ -1255,7 +1262,7 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
      *
      * @return A new traversable containing all elements of this traversable paired with their index, starting with 0.
      */
-    Traversable<Tuple2<T, Long>> zipWithIndex();
+    Traversable<Tuple2<T, Integer>> zipWithIndex();
 
     /**
      * Returns a traversable formed from this traversable and another Iterable collection by mapping elements.
@@ -1269,6 +1276,6 @@ public interface Traversable<T> extends Foldable<T>, Value<T> {
      * @return a new traversable containing mapped elements of this traversable and {@code that} iterable.
      * @throws NullPointerException if {@code mapper} is null
      */
-    <U> Traversable<U> zipWithIndex(BiFunction<? super T, ? super Long, ? extends U> mapper);
+    <U> Traversable<U> zipWithIndex(BiFunction<? super T, ? super Integer, ? extends U> mapper);
 
 }
